ML models refactoring
Split ML models into files fo readability, added return types to all functions, corrected linter warnings
