+++ gym/captchas/train.py
@@ -1,15 +1,16 @@
 # Imports
-import ast
 import torch
 from torch import nn
 import torch.optim as optim
-from torchvision import transforms, datasets
-import numpy as np
 from torch.utils.data import DataLoader
-import os
 import datetime
-from model.modelMulti import ModelMultiBaseline, ModelMultiResnet, ModelMultiSimple, MultiObjectDataset, train_multi
-from model.modelSingle import ModelSingle, ObjectDetectionDataset, train_single
+from torch.optim.adam import Adam
+from models.multi.dataset import MultiObjectDataset
+from models.multi.modelMulti import ModelMulti
+from models.multi.trainLoop import train_multi
+from models.single.dataset import SingleObjectDataset
+from models.single.modelSingle import ModelSingle
+from models.single.trainLoop import train_single
 
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 print(f"Using device: {device}")
@@ -52,12 +53,12 @@ class TrainerMulti:
         
 
     def train(self) -> None:
-        model = ModelMultiResnet(self.CLASS_COUNT).to(device)
+        model = ModelMulti(self.CLASS_COUNT).to(device)
         criterion = nn.BCELoss()
-        optimizer = optim.Adam(model.parameters(), lr=0.001)
+        optimizer = Adam(model.parameters(), lr=0.001)
         train_multi(model, criterion, optimizer, self.dataloaders, self.datasets, FREEZED_EPOCHS)
         model.unfreeze_last_resnet_layer()
-        optimizer = optim.Adam(model.parameters(), lr=0.0001)
+        optimizer = Adam(model.parameters(), lr=0.0001)
         history = train_multi(model, criterion, optimizer, self.dataloaders, self.datasets, EPOCHS)
         max_accuracy = int(max([value.cpu().item() for value in history['val_accuracy']]) * 100)
         torch.save(model.state_dict(), f'{CAPTCHA_RESULT_MODELS_DIR}/model_multi_{datetime.datetime.now().strftime("%Y-%m-%d_%H-%M")}_{max_accuracy}.pt')
@@ -67,10 +68,10 @@ class TrainerMulti:
 
 class TrainerSingle:
     def __init__(self) -> None:
-        self.datasets: dict[str, ObjectDetectionDataset] = {
-            'train': ObjectDetectionDataset(images_dir='input/yolo/train/images', labels_dir='input/yolo/train/labels', is_training=True),
-            'val': ObjectDetectionDataset(images_dir='input/yolo/valid/images', labels_dir='input/yolo/valid/labels', is_training=False),
-            'test': ObjectDetectionDataset(images_dir='input/yolo/test/images', labels_dir='input/yolo/test/labels', is_training=False)
+        self.datasets: dict[str, SingleObjectDataset] = {
+            'train': SingleObjectDataset(images_dir='input/yolo/train/images', labels_dir='input/yolo/train/labels', is_training=True),
+            'val': SingleObjectDataset(images_dir='input/yolo/valid/images', labels_dir='input/yolo/valid/labels', is_training=False),
+            'test': SingleObjectDataset(images_dir='input/yolo/test/images', labels_dir='input/yolo/test/labels', is_training=False)
         }
         
         self.dataloaders = {
@@ -84,14 +85,14 @@ class TrainerSingle:
     def train(self) -> None:
         model = ModelSingle(num_classes=self.NUM_CLASSES)
         criterion = nn.BCELoss()
-        optimizer = optim.Adam(model.parameters(), lr=0.001)
-        train_single(model, self.dataloaders, criterion, optimizer, FREEZED_EPOCHS, self.datasets)
+        optimizer = Adam(model.parameters(), lr=0.001)
+        train_single(model, criterion, optimizer, self.dataloaders,  self.datasets, FREEZED_EPOCHS)
         model.unfreeze_last_resnet_layer()
-        optimizer = optim.Adam(model.parameters(), lr=0.0001)
-        history = train_single(model, self.dataloaders, criterion, optimizer, UNFREEZED_LAST_LAYER_EPOCHS, self.datasets)
+        optimizer = Adam(model.parameters(), lr=0.0001)
+        history = train_single(model, criterion, optimizer, self.dataloaders, self.datasets, UNFREEZED_LAST_LAYER_EPOCHS)
         model.unfreeze_second_to_last_resnet_layer()
-        optimizer = optim.Adam(model.parameters(), lr=0.0001)
-        history = train_single(model, self.dataloaders, criterion, optimizer, EPOCHS, self.datasets)
+        optimizer = Adam(model.parameters(), lr=0.0001)
+        history = train_single(model, criterion, optimizer, self.dataloaders, self.datasets, EPOCHS)
         max_accuracy = int(max([value.cpu().item() for value in history['val_accuracy']]) * 100)
         torch.save(model.state_dict(), f'{CAPTCHA_RESULT_MODELS_DIR}/model_single_{datetime.datetime.now().strftime("%Y-%m-%d_%H-%M")}_{max_accuracy}.pt')
         model.plot_accuracy_from_history(history)  #, path="accuracy_plot.png")
@@ -101,6 +102,6 @@ class TrainerSingle:
 
 if __name__ == '__main__':
     trainer = TrainerMulti()
-    trainer.train()    
+    trainer.train()
     # trainer_single = TrainerSingle()
     # trainer_single.train()
